fixed op_live nondecreasing c_t_d bug with 21 live

+++ VM/srcs/get_arg.c
@@ -6,7 +6,7 @@
 /*   By: aelphias <aelphias@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/13 12:14:08 by aelphias          #+#    #+#             */
-/*   Updated: 2021/01/18 17:29:05 by aelphias         ###   ########.fr       */
+/*   Updated: 2021/01/18 18:27:10 by aelphias         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,9 +42,9 @@ int		read_int(const uint8_t *arena, int pos, int size)
 
 int		 get_arg(t_vm *vm, t_car *car, int arg_number, uint8_t *arena)
 {
-	t_op	*op;
-	int		value;
-	int		pos;
+	t_op		*op;
+	long int	value;
+	int			pos;
 
 	pos = 0;
 	value = 0;
